More work to get all spec tests at least parsing & compiling (#116)
* Handle the Memory* operands types in text parser
* Add support for TableInit/TableCopy instructions
* Fix ref.null parsing (heaptype, not reftype)
* Add parsing of memory & data module fields
* Fix handling of bare function idx list in elem list
* Set a max stack depth of 256
* Parse register cmds in spec tests properly
* Fix instruction data opcode for secondary instructions

+++ codegen/src/code.rs
@@ -4,7 +4,7 @@ use std::io::{Result, Write};
 
 /// Emit the execution functions for all of the [Instructions][Instruction] in the provided set.
 pub trait EmitCode: Write + std::fmt::Debug {
-    fn emit_code_file(&mut self, insts: &HashMap<u32, Instruction>) -> Result<()> {
+    fn emit_code_file(&mut self, insts: &HashMap<u8, Instruction>) -> Result<()> {
         self.write_all(CODE_HEADER.as_bytes())?;
 
         for (_, inst) in insts.iter() {
